Add dev dependencies for linter to enter /ui
It appears to be impossible to make the top-level lint config ignore
ui/.eslintrc.cjs; adding an ignore path for ui/** doesn't keep it from
reading the config in ui/ and failing on missing dev depdencies.

Instead, just add the dependencies needed by eslint to also check the
ui/ subdir.